@@ -26,7 +26,7 @@ install_requires =
     docutils
     pygments
     blessed
-python_requires = >=3.7
+python_requires = >=3.8
 
 [options.packages.find]
 exclude =
